fix(tests): move the suite out of the published directory, and run it in CI (#691)
marketing-site/ IS the Pages build output (pages_build_output_dir = "."), so
everything in it is uploaded by wrangler pages deploy. The tests lived there and
were publicly downloadable — 19KB of test source, and for a while an esbuild
bundle of issue.ts / present.ts / seats.ts / crypto.ts / auth.ts / jwt.ts as one
file (fixed separately in #695).

There is no exclusion mechanism to reach for. Pages honours neither .gitignore
nor dotfile prefixes — the old bundle was both and was published anyway — and
.assetsignore is a WORKERS static-assets feature, not a Pages one: the Cloudflare
docs describe it only for Workers, contrasting it with what 'Pages would
automatically exclude'. Being outside the directory is the only reliable answer.

So marketing-site/tests -> marketing-site-tests/. Two consequences handled:

  - Node resolves bare imports relative to the importing FILE, so run.mjs could
    no longer find esbuild. It now resolves esbuild from marketing-site's own
    node_modules via createRequire, anchored to run.mjs rather than cwd, instead
    of duplicating the dependency.
  - Test discovery is relative to run.mjs, not cwd, so it does not matter where
    npm invokes it from. cwd must still be marketing-site: the bundled code reads
    functions/api/schema.sql by a cwd-relative path and the bundle lives under
    marketing-site/node_modules so miniflare resolves.

Also adds .github/workflows/marketing-site.yml, because NO workflow ran these 34
tests. They executed only on a developer's machine, leaving the licence
issue/present path, the contact endpoint, the email ledger, trial-expiry
entitlement and the seat caps unguarded on merge — the same 'counted but never
executed' failure CLAUDE.md 3 records for the Clash and Routing projects.

The workflow ends with a guard that fails if any *.test.ts reappears inside
marketing-site/, so this cannot regress silently. Both directions were exercised
locally: clean tree passes, and a planted functions/oops.test.ts trips it.

Verified: 34 tests pass from the new location; a deliberately failing file exits
1 and a clean run exits 0; workflow YAML parses.

Note tsconfig's include is functions/**/*.ts, so the tests were NEVER
typechecked — before or after this move. Unchanged here; covering them needs
@types/node and is separate.
